implement the interfaces expected by dogfood
